Fixed identifier lexer, made lexer support comments, made integer bigint, refactoring

+++ mini-pl/src/lexer/tokens.rs
@@ -1,3 +1,5 @@
+use num_bigint::BigInt;
+
 use Ident;
 
 #[derive(Clone, Debug, PartialEq)]
@@ -11,7 +13,7 @@ pub enum Token {
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Literal {
-    Integer(i64),
+    Integer(BigInt),
     StringLit(String),
 }
